+++ src/components/index.ts
@@ -28,6 +28,9 @@ export * from "./paginator";
 export { default as Search } from "./search";
 export * from "./search";
 
+export { default as Stepper } from "./stepper";
+export * from "./stepper";
+
 export { default as Table } from "./table";
 export * from "./table";
 
@@ -36,6 +39,9 @@ export * from "./tabs";
 
 export { default as Tag } from "./tag";
 export * from "./tag";
+
+export { default as Timeline } from "./timeline";
+export * from "./timeline";
 
 // Progress/Loader component
 export { default as Progress } from "./progress";
@@ -59,6 +65,15 @@ export * from "./primitives/number";
 
 export { default as Password } from "./primitives/password";
 export * from "./primitives/password";
+
+export { default as RadioGroup } from "./primitives/radioGroup";
+export * from "./primitives/radioGroup";
+
+export { default as Slider } from "./primitives/slider";
+export * from "./primitives/slider";
+
+export { default as Switch } from "./primitives/switch";
+export * from "./primitives/switch";
 
 export { default as Text } from "./primitives/text";
 export * from "./primitives/text";